+++ src/GitLabApiClient/Models/Notes/Responses/Note.cs
@@ -5,10 +5,10 @@ namespace GitLabApiClient.Models.Notes.Responses
     public sealed class Note : ModifiableObject
     {
         [JsonProperty("noteable_id")]
-        public int NoteableId { get; set; }
+        public long NoteableId { get; set; }
 
         [JsonProperty("noteable_iid")]
-        public int NoteableIid { get; set; }
+        public long NoteableIid { get; set; }
 
         [JsonProperty("noteable_type")]
         public string NoteableType { get; set; }
